requirements.txt
flask
flask-cors
pillow
matplotlib
opencv-python
torch
torchvision